Revert "Move service container binding tests to menu test"
This reverts commit c9b1127f3cd6d31c319c77c04bc3b152e3a7bcb8.
